+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XInput/XInputControllerWindows.cs
@@ -84,7 +84,7 @@ namespace UnityEngine.Experimental.Input.Plugins.XInput.LowLevel
 
         public XInputControllerWindowsState WithButton(Button button)
         {
-            buttons |= (uint)1 << (int)button;
+            buttons |= (ushort)((uint)1 << (int)button);
             return this;
         }
     }
